backend/scripts/requirements.txt
google-cloud-vision>=3.5.0
pdf2image>=1.16.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
mysql-connector-python>=8.2.0
